requirements.txt
openai
mutagen
elevenlabs==0.3.0b0
moviepy
pydub
nltk
pandas
Pillow
sympy
matplotlib
gspread
google-auth
numpy
